Configure smtpd.conf for SASL and add an postfix_sasldb_user LWRP
This allows Postfix to authenticate clients using SASL, using either a
sasldb file or saslauthd. The presence of the existing sasl_auth
recipe for using SASL as a client makes the situation a little
confusing. Perhaps a less ambiguous name would have helped.

It was tricky to figure out where smtpd.conf should be
configured. CentOS installs it along with Postfix itself but other
platforms don't seem to install it at all. It's not much use without
the SASL packages so I lumped it into the new sasl_common recipe. It
does no harm to those not wanting to authenticate clients with SASL.

The default attributes for configuring smtpd.conf mirror the CentOS
defaults, which is to use PLAIN or LOGIN against saslauthd, which in
turn uses PAM. These seem like good defaults.

The sasldb recipe was added primarily for testing as most will
probably want to use the LWRP directly in conjunction with encrypted
data bags or something but it may be useful to someone. :)

Signed-off-by: James Le Cuirot <james.le-cuirot@yakara.com>