Fix OpenSUSE installation issues
OpenSUSE runs a script which builds the main.cf file automatically if a file
doesn't exist. This breaks idempotency so go ahead and touch the file via Chef.

In addition, postfix is built with incorrect default settings for `setgid_group`
and `daemon_directory`, so set it properly in the config.

Signed-off-by: Lance Albertson <lance@osuosl.org>